feat: add README.md

+++ src/main/java/com/example/schedule/service/UserService.java
@@ -52,18 +52,28 @@ public class UserService {
         return new UserResponseDto(findUser.getUsername(),findUser.getEmail() , findUser.getPassword());
     }
 
-    public void delete(Long id){
+    public void delete(Long id , String username){
 
+        User loginUser = userRepository.findUserByUsernameOrElseThrow(username);
         User findUser = userRepository.findByIdOrElseThrow(id);
-
+        //로그인한 유저와 삭제하려는 유저 id 일치하는지 판별
+        if (loginUser.getId() != findUser.getId()){
+              throw new CustomException(UNAUTHORIZED_USER);
+        }
         userRepository.delete(findUser);
     }
 
     public LoginResponseDto findUserByEmailAndPasswordOrElseThrow(String email , String password){
 
         Optional<User> optionalUser = userRepository.findUserByEmail(email);
+
+        // email 에 일치하는 유저 판별
+        if (optionalUser.isEmpty()){
+            throw new CustomException(UNAUTHORIZED_EMAIL);
+        }
         User findUser = optionalUser.get();
 
+        // 해당 유저 password 판별
         if (!passwordEncoder.matches(password , findUser.getPassword())){
             throw new CustomException(UNAUTHORIZED_PASSWORD);
         }
